roles&permission

+++ app/Modules/Master/routes.php
@@ -11,29 +11,29 @@ use App\Modules\Master\Controllers\CityController;
 use Illuminate\Support\Facades\Route;
 
 // Department
-Route::get('/department', [DepartmentController::class, 'index'])->name('department.index');
-Route::post('/department/store', [DepartmentController::class, 'store'])->name('department.store');
-Route::put('/department/{id}', [DepartmentController::class, 'update'])->name('department.update');
-Route::delete('/department/{id}', [DepartmentController::class, 'destroy'])->name('department.destroy');
+Route::middleware('check.permission:read,department')->get('/department', [DepartmentController::class, 'index'])->name('department.index');
+Route::middleware('check.permission:create,department')->post('/department/store', [DepartmentController::class, 'store'])->name('department.store');
+Route::middleware('check.permission:edit,department')->put('/department/{id}', [DepartmentController::class, 'update'])->name('department.update');
+Route::middleware('check.permission:delete,department')->delete('/department/{id}', [DepartmentController::class, 'destroy'])->name('department.destroy');
 
 // Designation
-Route::get('/designation', [DesignationController::class, 'index'])->name('designation.index');
-Route::post('/designation/store', [DesignationController::class, 'store'])->name('designation.store');
-Route::put('/designation/{id}', [DesignationController::class, 'update'])->name('designation.update');
-Route::delete('/designation/{id}', [DesignationController::class, 'destroy'])->name('designation.destroy');
-Route::get('/designation/export/excel', [DesignationController::class, 'exportExcel'])->name('designation.export.excel');
+Route::middleware('check.permission:read,designation')->get('/designation', [DesignationController::class, 'index'])->name('designation.index');
+Route::middleware('check.permission:create,designation')->post('/designation/store', [DesignationController::class, 'store'])->name('designation.store');
+Route::middleware('check.permission:edit,designation')->put('/designation/{id}', [DesignationController::class, 'update'])->name('designation.update');
+Route::middleware('check.permission:delete,designation')->delete('/designation/{id}', [DesignationController::class, 'destroy'])->name('designation.destroy');
+Route::middleware('check.permission:read,designation')->get('/designation/export/excel', [DesignationController::class, 'exportExcel'])->name('designation.export.excel');
 
 // Branch
-Route::get('/new-branch', [BranchController::class, 'index'])->name('new-branch.index');
-Route::post('/new-branch/store', [BranchController::class, 'store'])->name('new-branch.store');
-Route::put('/new-branch/{id}', [BranchController::class, 'update'])->name('new-branch.update');
-Route::delete('/new-branch/{id}', [BranchController::class, 'destroy'])->name('new-branch.destroy');
+Route::middleware('check.permission:read,branch')->get('/new-branch', [BranchController::class, 'index'])->name('new-branch.index');
+Route::middleware('check.permission:create,branch')->post('/new-branch/store', [BranchController::class, 'store'])->name('new-branch.store');
+Route::middleware('check.permission:edit,branch')->put('/new-branch/{id}', [BranchController::class, 'update'])->name('new-branch.update');
+Route::middleware('check.permission:delete,branch')->delete('/new-branch/{id}', [BranchController::class, 'destroy'])->name('new-branch.destroy');
 
 // Location
-Route::get('/location', [LocationController::class, 'index'])->name('location.index');
-Route::post('/location/store', [LocationController::class, 'store'])->name('location.store');
-Route::put('/location/{id}', [LocationController::class, 'update'])->name('location.update');
-Route::delete('/location/{id}', [LocationController::class, 'destroy'])->name('location.destroy');
+Route::middleware('check.permission:read,location')->get('/location', [LocationController::class, 'index'])->name('location.index');
+Route::middleware('check.permission:create,location')->post('/location/store', [LocationController::class, 'store'])->name('location.store');
+Route::middleware('check.permission:edit,location')->put('/location/{id}', [LocationController::class, 'update'])->name('location.update');
+Route::middleware('check.permission:delete,location')->delete('/location/{id}', [LocationController::class, 'destroy'])->name('location.destroy');
 
 // Country
 Route::get('/country', [CountryController::class, 'index'])->name('country.index');
@@ -58,3 +58,6 @@ Route::get('/city', [CityController::class, 'index'])->name('city.index');
 Route::post('/city/store', [CityController::class, 'store'])->name('city.store');
 Route::put('/city/{id}', [CityController::class, 'update'])->name('city.update');
 Route::delete('/city/{id}', [CityController::class, 'destroy'])->name('city.destroy');
+
+// RBAC Management
+require __DIR__ . '/rbac-routes.php';
